Adjust smtpsink to use portguard
